requirements.txt
matplotlib==3.7.2
numpy==1.24.1
Pillow==10.0.1
torch==2.0.1+cu118
torchsummary==1.5.1
torchvision==0.15.2+cu118
tqdm==4.66.1
